Add commands to activationEvents explicitly
You'd think that this shouldn't happen with the wildcard `"*"` `activationEvent`, but I'm often greeted by errors like this when running a project manager command _right after_ opening a VSCode window (so when PM hasn't been loaded yet presumably):

![](http://i.imgur.com/L59zKVT.png)

Adding the command(s) as `activationEvents` explicitly seems to fix this (although there is a slight delay until the quick pick shows up, VSCode probably has to load PM first). Not sure if this is working as designed or a bug in VSCode, but it seems like an OK workaround.