Add ability to temporarily disable individual projects (#154)
This is an initial implementation that requires manual editing of the
config. Future enchancements could include commands to enable/disable
the current (or a selected) project.

It maintains backwards compatibity with the existing storage format by
setting the default value to `true`.  It also provides default values
for all other properties of a `Project`, but they may not *really* be
valid (e.g. an empty string for name). It could be improved by actually
validating the loaded data and handling invalid entries (just ignore
them, probably?).